// web/src/api/types.ts
export interface components {
    schemas: {
// ...
        GrantOut: {
// ...
            status: string;
        };
// ...
        RegisterAnalysisIn: {
            /** Object Key */
            object_key: string;
            analysis_type: components["schemas"]["AnalysisType"];
            /** Label */
            label: string;
            /** Lab Date */
            lab_date?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}